Fix some problems

+++ src/stores/score.js
@@ -4,7 +4,7 @@ export const score = reactive({
   score: parseInt(localStorage.getItem('scoreCurrentQuiz')) || 0,
   profileScore: parseInt(localStorage.getItem('profileScore')) || 0,
   questionSucess: JSON.parse(localStorage.getItem('questionsSucessfull')) || [],
-  expressLearned: JSON.parse(localStorage.getItem('numberOfExpressMaterised')) || 0 ,
+  expressLearned: JSON.parse(localStorage.getItem('numberOfExpressMaterised')) || 0,
 
   increment() {
     this.score++;
@@ -25,13 +25,22 @@ export const score = reactive({
 
 
   goodAnswer(questionSucess) {
-    this.questionSucess.push(questionSucess);
 
-    this.expressLearned = this.questionSucess.length
+    let storeQuestionsSucessfull = JSON.parse(localStorage.getItem('questionsSucessfull')) || [];
 
-    localStorage.setItem('questionsSucessfull', JSON.stringify(this.questionSucess));
-    localStorage.setItem('numberOfExpressMaterised', this.expressLearned);
-    console.log('Nouvelle question ajoutée dans le store:', questionSucess);
-    console.log('Nombre d\'expressions maîtrisées:', this.expressLearned);
+    const alreadyAdded = storeQuestionsSucessfull.some(q => JSON.stringify(q) === JSON.stringify(questionSucess));
+
+    if (!alreadyAdded) {
+      this.questionSucess.push(questionSucess);
+
+      this.expressLearned = this.questionSucess.length
+
+      localStorage.setItem('questionsSucessfull', JSON.stringify(this.questionSucess));
+      localStorage.setItem('numberOfExpressMaterised', this.expressLearned);
+      // console.log('Nouvelle question ajoutée dans le store:', questionSucess);
+      // console.log('Nombre d\'expressions maîtrisées:', this.expressLearned);
+    }else {
+      console.log("Déjà dans le localstorage")
+    }
   }
 });
